+++ radio2/src/loc.h
@@ -1,4 +1,4 @@
-// $Id: loc.h,v 1.6 2009-11-15 23:33:54 nick Exp $
+// $Id: loc.h,v 1.7 2009-11-16 06:40:33 nick Exp $
 
 #ifndef _LOC_H
 #define _LOC_H
@@ -12,7 +12,7 @@ typedef struct loc_s {
 } __attribute__((__packed__)) loc_t;
 
 unsigned long loc_dist2(loc_t *a, loc_t *b);
-float loc_dist(loc_t *a, loc_t *b);
+double loc_dist(loc_t *a, loc_t *b);
 void loc_zero(loc_t *a);
 void loc_perturb(loc_t *a, unsigned int d);
 void loc_fprint(loc_t *a, FILE *fp);
